rename to metadata.json
