+++ .github/workflows/update-car-metadata.yml
@@ -38,7 +38,7 @@ jobs:
         run: |
           cd opendbc
           python ../scripts/get_metadata.py
-          mv metadata.json ../src/data/car_data.json
+          mv metadata.json ../src/data/metadata.json
 
       - name: Create Pull Request
         uses: peter-evans/create-pull-request@af7c021bb9c41362d0986422bd244ce51a3bf52f
