+++ src/pages/cars/[car].astro
@@ -1,6 +1,6 @@
 ---
 import BaseLayout from '../../layouts/BaseLayout.astro';
-import cars from '../../data/car_data.json';
+import cars from '../../data/metadata.json';
 import { getSupportTypeColor } from '../../utils/supportType';
 import InfoCard from '../../components/InfoCard';
 import MasterToggle from '../../components/MasterToggle';
