+++ src/pages/index.astro
@@ -1,7 +1,7 @@
 ---
 import BaseLayout from '../layouts/BaseLayout.astro';
 import Layout from '../components/Layout';
-import cars from '../data/car_data.json';
+import cars from '../data/metadata.json';
 ---
 
 <BaseLayout title="opendbc">
